Create calling methods using Object
Create a class Flight to store the details of flight number, destination, distance of the flight, fuel
required by the flight. Write methods to enter the details of the flight and to calculate the value of fuel
as per the following criteria: Distance<=600, value of Fuel=300, for next 1000 km, value of fuel=500,
greater than 1000 distance, fuel=400. Write the main method to create an object of the class and call
the member methods.